requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pystray==0.19.5
six==1.17.0
